scripts/requirements.txt
psycopg2-binary>=2.9.3
requests>=2.28.1
redis